re-add success toast logic
